Added debug layer output and  fixed issue with kernel offset types

+++ src/layers/anime4k/conv2d-3x4.ts
@@ -49,11 +49,11 @@ class Anime4KConv3x4 extends ComputeLayer {
                 let i = id.y*256 + x;
                 var result  = vec4f(0.0, 0.0, 0.0, 0.0);
                 
-                let coord = vec2<u32>(x,y);
+                let coord = vec2<i32>( i32(x), i32(y));
                       
                  for(var i = 0u; i < 9; i++){
-                   let offset = vec2<u32>(kernel_offsets[i].xy);
-                   result += kernels[i]*textureLoad(inputTexture0, coord+offset, 0);
+                   let offset = vec2<i32>(kernel_offsets[i].xy);
+                   result += kernels[i]*textureLoad(inputTexture0, coord + offset, 0);
                  } 
                     
                 result += bias;
